+++ performance_test/src/experiment_metrics/analysis_result.hpp
@@ -15,16 +15,22 @@
 #ifndef EXPERIMENT_METRICS__ANALYSIS_RESULT_HPP_
 #define EXPERIMENT_METRICS__ANALYSIS_RESULT_HPP_
 
+#include <chrono>
+#include <sstream>
+#include <string>
+
 #if !defined(WIN32)
 #include <sys/time.h>
 #include <sys/resource.h>
 #endif  // !defined(WIN32)
 
-#include <chrono>
-#include <sstream>
-#include <string>
+#if defined(QNX)
+#include <sys/neutrino.h>
+#include <sys/syspage.h>
+#include <sys/types.h>
+#endif  // defined(QNX)
 
-#include "../utilities/statistics_tracker.hpp"
+#include "../utilities/sample_statistics.hpp"
 #include "../utilities/cpu_usage_tracker.hpp"
 
 namespace performance_test
@@ -39,6 +45,12 @@ struct AnalysisResult
   AnalysisResult();
   static std::string csv_header(const bool pretty_print = false, std::string st = ",");
   std::string to_csv_string(const bool pretty_print = false, std::string st = ",") const;
+  size_t latency_seconds_n() const;
+  double latency_seconds_min() const;
+  double latency_seconds_max() const;
+  double latency_seconds_mean() const;
+  double latency_seconds_m2() const;
+  double latency_seconds_variance() const;
 
   std::chrono::nanoseconds m_experiment_start = {};
   std::chrono::nanoseconds m_time_between_two_measurements = {};
@@ -47,7 +59,7 @@ struct AnalysisResult
   uint64_t m_num_samples_lost = {};
   std::size_t m_total_data_received = {};
 
-  StatisticsTracker m_latency;
+  SampleStatistics<std::int64_t> m_latency_stats;
 #if !defined(WIN32)
   rusage m_sys_usage;
 #endif  // !defined(WIN32)